// foodguard-backend/src/modules/users/domain/enums/user-status.enum.ts
export enum UserRole {
  DONOR = 'DONOR',
  INDIVIDUAL_RECIPIENT = 'INDIVIDUAL_RECIPIENT',
  ORGANIZATIONAL_RECIPIENT = 'ORGANIZATIONAL_RECIPIENT',
  ADMIN = 'ADMIN',
}

export enum VerificationStatus {
  UNVERIFIED = 'UNVERIFIED',
  PENDING = 'PENDING',
  VERIFIED = 'VERIFIED',
  REJECTED = 'REJECTED',
}